% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche.runif.R
\name{niche.runif}
\alias{niche.runif}
\title{Uniform sampling from elliptical niche region.}
\usage{
niche.runif(n, mu, Sigma, alpha = 0.95)
}
\arguments{
\item{n}{number of random draws}

\item{mu}{mean vector}

\item{Sigma}{variance matrix}

\item{alpha}{probabilistic niche size}
}
\value{
iid draws from uniform distribution on niche region.
}
\description{
Uniform sampling from elliptical niche region.
}
\examples{
# 2d example
d <- 2 # number of dimensions
V <- crossprod(matrix(rnorm(4),d,d))
mu <- rnorm(d)
plot(ellipse(mu, V), type = "l")
points(niche.runif(1e4, mu, V), col = "brown", pch = ".")
}
